// packages/angular/src/executors/file-server/schema.d.ts
export interface Schema {
  host: string;
  port: number;
  ssl: boolean;
  sslKey?: string;
  sslCert?: string;
  proxyUrl?: string;
  buildTarget: string;
  parallel: boolean;
  maxParallel?: number;
  withDeps: boolean;
  proxyOptions?: object;
  watch?: boolean;
  spa?: boolean;
  cors?: boolean;
}